moved authviewmodel
